+++ server/filesystem/disk_space.go
@@ -1,12 +1,14 @@
 package filesystem
 
 import (
+	"slices"
 	"sync"
 	"sync/atomic"
 	"time"
 
 	"emperror.dev/errors"
 	"github.com/apex/log"
+	"golang.org/x/sys/unix"
 
 	"github.com/mythicalltd/featherwings/internal/ufs"
 )
@@ -164,7 +166,9 @@ func (fs *Filesystem) DirectorySize(root string) (int64, error) {
 		return 0, err
 	}
 
+	var hardLinks []uint64
 	var size atomic.Int64
+
 	err = fs.unixFS.WalkDirat(dirfd, name, func(dirfd int, name, _ string, d ufs.DirEntry, err error) error {
 		if err != nil {
 			return errors.Wrap(err, "walkdirat err")
@@ -180,8 +184,16 @@ func (fs *Filesystem) DirectorySize(root string) (int64, error) {
 			return errors.Wrap(err, "lstatat err")
 		}
 
-		// TODO: detect if info is a hard-link and de-duplicate it.
-		// ref; https://github.com/mythicalltd/featherwings/pull/181/files
+		var sysFileInfo = info.Sys().(*unix.Stat_t)
+		if sysFileInfo.Nlink > 1 {
+			// Hard links have the same inode number
+			if slices.Contains(hardLinks, sysFileInfo.Ino) {
+				// Don't add hard links size twice
+				return nil
+			} else {
+				hardLinks = append(hardLinks, sysFileInfo.Ino)
+			}
+		}
 
 		size.Add(info.Size())
 		return nil